fix: check covenants still valid from previous Staking param for delegation expansion
