+++ src/staking/observable/index.ts
@@ -3,7 +3,8 @@ import { UTXO } from "../../types/UTXO";
 import { StakingError, StakingErrorCode } from "../../error";
 import { stakingTransaction } from "../transactions";
 import { isTaproot } from "../../utils/btc";
-import { toBuffers, validateStakingTxInputData } from "../../utils/staking";
+import { toBuffers } from "../../utils/staking";
+import { validateStakingTxInputData } from "../../utils/staking/validation";
 import { TransactionResult } from "../../types/transaction";
 import { ObservableStakingScriptData, ObservableStakingScripts } from "./observableStakingScript";
 import { StakerInfo, Staking } from "..";
