+++ src/utils/staking/index.ts
@@ -1,13 +1,8 @@
 import { address, networks, payments, Transaction } from "bitcoinjs-lib";
 import { Taptree } from "bitcoinjs-lib/src/types";
 import { internalPubkey } from "../../constants/internalPubkey";
-import { MIN_UNBONDING_OUTPUT_VALUE } from "../../constants/unbonding";
 import { StakingError, StakingErrorCode } from "../../error";
-import { StakingParams } from "../../types/params";
 import { TransactionOutput } from "../../types/psbtOutputs";
-import { UTXO } from "../../types/UTXO";
-import { isValidNoCoordPublicKey } from "../btc";
-
 export interface OutputInfo {
   scriptPubKey: Buffer;
   outputAddress: string;
@@ -204,176 +199,6 @@ export const findMatchingTxOutputIndex = (
 };
 
 /**
- * Validate the staking transaction input data.
- *
- * @param {number} stakingAmountSat - The staking amount in satoshis.
- * @param {number} timelock - The staking time in blocks.
- * @param {StakingParams} params - The staking parameters.
- * @param {UTXO[]} inputUTXOs - The input UTXOs.
- * @param {number} feeRate - The Bitcoin fee rate in sat/vbyte
- * @throws {StakingError} - If the input data is invalid.
- */
-export const validateStakingTxInputData = (
-  stakingAmountSat: number,
-  timelock: number,
-  params: StakingParams,
-  inputUTXOs: UTXO[],
-  feeRate: number,
-) => {
-  if (
-    stakingAmountSat < params.minStakingAmountSat ||
-    stakingAmountSat > params.maxStakingAmountSat
-  ) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_INPUT,
-      "Invalid staking amount",
-    );
-  }
-
-  if (
-    timelock < params.minStakingTimeBlocks ||
-    timelock > params.maxStakingTimeBlocks
-  ) {
-    throw new StakingError(StakingErrorCode.INVALID_INPUT, "Invalid timelock");
-  }
-
-  if (inputUTXOs.length == 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_INPUT,
-      "No input UTXOs provided",
-    );
-  }
-  if (feeRate <= 0) {
-    throw new StakingError(StakingErrorCode.INVALID_INPUT, "Invalid fee rate");
-  }
-};
-
-/**
- * Validate the staking parameters.
- * Extend this method to add additional validation for staking parameters based
- * on the staking type.
- * @param {StakingParams} params - The staking parameters.
- * @throws {StakingError} - If the parameters are invalid.
- */
-export const validateParams = (params: StakingParams) => {
-  // Check covenant public keys
-  if (params.covenantNoCoordPks.length == 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Could not find any covenant public keys",
-    );
-  }
-  if (params.covenantNoCoordPks.length < params.covenantQuorum) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Covenant public keys must be greater than or equal to the quorum",
-    );
-  }
-  params.covenantNoCoordPks.forEach((pk) => {
-    if (!isValidNoCoordPublicKey(pk)) {
-      throw new StakingError(
-        StakingErrorCode.INVALID_PARAMS,
-        "Covenant public key should contains no coordinate",
-      );
-    }
-  });
-  // Check other parameters
-  if (params.unbondingTime <= 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Unbonding time must be greater than 0",
-    );
-  }
-  if (params.unbondingFeeSat <= 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Unbonding fee must be greater than 0",
-    );
-  }
-  if (params.maxStakingAmountSat < params.minStakingAmountSat) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Max staking amount must be greater or equal to min staking amount",
-    );
-  }
-  if (
-    params.minStakingAmountSat <
-    params.unbondingFeeSat + MIN_UNBONDING_OUTPUT_VALUE
-  ) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      `Min staking amount must be greater than unbonding fee plus ${MIN_UNBONDING_OUTPUT_VALUE}`,
-    );
-  }
-  if (params.maxStakingTimeBlocks < params.minStakingTimeBlocks) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Max staking time must be greater or equal to min staking time",
-    );
-  }
-  if (params.minStakingTimeBlocks <= 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Min staking time must be greater than 0",
-    );
-  }
-  if (params.covenantQuorum <= 0) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_PARAMS,
-      "Covenant quorum must be greater than 0",
-    );
-  }
-  if (params.slashing) {
-    if (params.slashing.slashingRate <= 0) {
-      throw new StakingError(
-        StakingErrorCode.INVALID_PARAMS,
-        "Slashing rate must be greater than 0",
-      );
-    }
-    if (params.slashing.slashingRate > 1) {
-      throw new StakingError(
-        StakingErrorCode.INVALID_PARAMS,
-        "Slashing rate must be less or equal to 1",
-      );
-    }
-    if (params.slashing.slashingPkScriptHex.length == 0) {
-      throw new StakingError(
-        StakingErrorCode.INVALID_PARAMS,
-        "Slashing public key script is missing",
-      );
-    }
-    if (params.slashing.minSlashingTxFeeSat <= 0) {
-      throw new StakingError(
-        StakingErrorCode.INVALID_PARAMS,
-        "Minimum slashing transaction fee must be greater than 0",
-      );
-    }
-  }
-};
-
-/**
- * Validate the staking timelock.
- *
- * @param {number} stakingTimelock - The staking timelock.
- * @param {StakingParams} params - The staking parameters.
- * @throws {StakingError} - If the staking timelock is invalid.
- */
-export const validateStakingTimelock = (
-  stakingTimelock: number,
-  params: StakingParams,
-) => {
-  if (
-    stakingTimelock < params.minStakingTimeBlocks ||
-    stakingTimelock > params.maxStakingTimeBlocks
-  ) {
-    throw new StakingError(
-      StakingErrorCode.INVALID_INPUT,
-      "Staking transaction timelock is out of range",
-    );
-  }
-};
-
-/**
  * toBuffers converts an array of strings to an array of buffers.
  *
  * @param {string[]} inputs - The input strings.
@@ -390,4 +215,62 @@ export const toBuffers = (inputs: string[]): Buffer[] => {
       "Cannot convert values to buffers",
     );
   }
+};
+
+
+/**
+ * Strips all signatures from a transaction by clearing both the script and
+ * witness data. This is due to the fact that we only need the raw unsigned
+ * transaction structure. The signatures are sent in a separate protobuf field
+ * when creating the delegation message in the Babylon.
+ * @param tx - The transaction to strip signatures from
+ * @returns A copy of the transaction with all signatures removed
+ */
+export const clearTxSignatures = (tx: Transaction): Transaction => {
+  tx.ins.forEach((input) => {
+    input.script = Buffer.alloc(0);
+    input.witness = [];
+  });
+  return tx;
+};
+
+/**
+ * Derives the merkle proof from the list of hex strings. Note the
+ * sibling hashes are reversed from hex before concatenation.
+ * @param merkle - The merkle proof hex strings.
+ * @returns The merkle proof in hex string format.
+ */
+export const deriveMerkleProof = (merkle: string[]) => {
+  const proofHex = merkle.reduce((acc: string, m: string) => {
+    return acc + Buffer.from(m, "hex").reverse().toString("hex");
+  }, "");
+  return proofHex;
+};
+
+/**
+ * Extracts the first valid Schnorr signature from a signed transaction.
+ *
+ * Since we only handle transactions with a single input and request a signature
+ * for one public key, there can be at most one signature from the Bitcoin node.
+ * A valid Schnorr signature is exactly 64 bytes in length.
+ *
+ * @param singedTransaction - The signed Bitcoin transaction to extract the signature from
+ * @returns The first valid 64-byte Schnorr signature found in the transaction witness data,
+ *          or undefined if no valid signature exists
+ */
+export const extractFirstSchnorrSignatureFromTransaction = (
+  singedTransaction: Transaction,
+): Buffer | undefined => {
+  // Loop through each input to extract the witness signature
+  for (const input of singedTransaction.ins) {
+    if (input.witness && input.witness.length > 0) {
+      const schnorrSignature = input.witness[0];
+
+      // Check that it's a 64-byte Schnorr signature
+      if (schnorrSignature.length === 64) {
+        return schnorrSignature; // Return the first valid signature found
+      }
+    }
+  }
+  return undefined;
 };
